refactor: move InstanceIdVec impl into standalone file
